modules/network-scanner/requirements.txt
requests~=2.31.0
psutil~=5.9.5
PyAutoGUI~=0.9.54
Pillow~=9.5.0
netifaces~=0.11.0
cryptography~=41.0.1
netaddr~=0.8.0
python-nmap~=0.7.1